Move hints and attributes into the right folder
