+++ MvvmCross.Forms/Presenters/IMvxFormsPagePresenter.cs
@@ -6,6 +6,7 @@ using System;
 using MvvmCross.Forms.Core;
 using MvvmCross.ViewModels;
 using MvvmCross.Presenters;
+using MvvmCross.Presenters.Attributes;
 using Xamarin.Forms;
 
 namespace MvvmCross.Forms.Presenters
