+++ MvvmCross/Platform/Android/Presenters/Attributes/MvxActivityPresentationAttribute.cs
@@ -7,6 +7,7 @@ using System.Collections.Generic;
 using Android.OS;
 using Android.Views;
 using MvvmCross.Presenters;
+using MvvmCross.Presenters.Attributes;
 
 namespace MvvmCross.Platform.Android.Presenters.Attributes
 {
