+++ MvvmCross/Platform/Wpf/Presenters/MvxSimpleWpfViewPresenter.cs
@@ -9,7 +9,6 @@ using System.Linq;
 using MvvmCross.Logging;
 using MvvmCross.ViewModels;
 using MvvmCross.Platform.Wpf.Views;
-using MvvmCross.ViewModels.Hints;
 
 namespace MvvmCross.Platform.Wpf.Presenters
 {
