+++ MvvmCross/Presenters/IMvxOverridePresentationAttribute.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MS-PL license.
 // See the LICENSE file in the project root for more information.
 
+using MvvmCross.Presenters.Attributes;
 using MvvmCross.ViewModels;
 
 namespace MvvmCross.Presenters
