+++ MvvmCross/Presenters/MvxPresentationAttributeAction.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
+using MvvmCross.Presenters.Attributes;
 using MvvmCross.ViewModels;
 
 namespace MvvmCross.Presenters
