+++ UnitTests/MvvmCross.UnitTest/Navigation/NavigationServiceTests.cs
@@ -11,7 +11,6 @@ using Moq;
  using MvvmCross.UnitTest.Mocks.Dispatchers;
  using MvvmCross.UnitTest.Mocks.ViewModels;
  using MvvmCross.ViewModels;
-using MvvmCross.ViewModels.Hints;
 using Xunit;
 
 namespace MvvmCross.UnitTest.Navigation
